Rename Header component to SiteHeader to better define its intended usage.
