--- src/app/components/header/header.stories.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { Header } from './header';
-
-storiesOf('Header', module)
-  .add('Page Header', () => <Header actions={{}}/>);
