+++ src/app/components/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-named-as-default */
-import Header from './header/header';
+import Header from './site-header/site-header';
 import Footer from './footer/footer';
 
 export default {
