+++ src/app/components/site-header/site-header.scss
