+++ src/app/components/site-header/site-header.stories.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import { SiteHeader } from './site-header';
+
+storiesOf('SiteHeader', module)
+  .add('Site Header', () => <SiteHeader actions={{}}/>);
